Organize solver options (solver is selectable by string)
